Feature user-friendly error messages

+++ src/pages/WelcomePage/WelcomePage.tsx
@@ -2,13 +2,13 @@ import React, { FC, useEffect } from 'react';
 import { ToastContainer, toast } from 'react-toastify';
 import { useAppSelector, useAppDispatch } from '~/hooks/redux';
 import { useTranslation } from 'react-i18next';
-import { resetIsDeleted } from '~/store/reducers/authSlice';
+import { resetIsDeleted, clearError } from '~/store/reducers/authSlice';
 import Footer from '~/components/Footer';
 
 import styles from './WelcomePage.module.scss';
 
 const WelcomePage: FC = () => {
-  const { isDeleted } = useAppSelector(state => state.auth);
+  const { isDeleted, error } = useAppSelector(state => state.auth);
   const { t } = useTranslation();
   const dispatch = useAppDispatch();
 
@@ -18,9 +18,20 @@ const WelcomePage: FC = () => {
         position: toast.POSITION.BOTTOM_RIGHT,
       });
     return () => {
-      dispatch(resetIsDeleted());
+      isDeleted && dispatch(resetIsDeleted());
     };
   }, [dispatch, isDeleted, t]);
+
+  useEffect(() => {
+    setTimeout(() => {
+      if (error) {
+        toast.error(error, {
+          position: toast.POSITION.BOTTOM_RIGHT,
+        });
+        dispatch(clearError());
+      }
+    }, 500);
+  }, [dispatch, error]);
 
   return (
     <>
